backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
chromadb
langchain
langchain-openai
python-multipart
pypdf